fix remote expression in TF
also:
* split synth.go into acl.go and sg.go
