+++ test/data/sg_testing2/sg_expected.tf
@@ -1,54 +1,54 @@
 # Internal. required-connections[2]: (instance fe-ky)->(instance be-ky); allowed-protocols[0]
 resource "ibm_is_security_group_rule" "sgrule-be-ky-0" {
-  group     = be-ky.id
+  group     = ibm_is_security_group.be-ky.id
   direction = "inbound"
-  remote    = fe-ky.id
+  remote    = ibm_is_security_group.fe-ky.id
   tcp {
   }
 }
 # Internal. required-connections[3]: (instance be-ky)->(instance opa-ky); allowed-protocols[0]
 resource "ibm_is_security_group_rule" "sgrule-be-ky-1" {
-  group     = be-ky.id
+  group     = ibm_is_security_group.be-ky.id
   direction = "outbound"
-  remote    = opa-ky.id
+  remote    = ibm_is_security_group.opa-ky.id
   tcp {
   }
 }
 # Internal. required-connections[1]: (instance proxy-ky)->(instance fe-ky); allowed-protocols[0]
 resource "ibm_is_security_group_rule" "sgrule-fe-ky-0" {
-  group     = fe-ky.id
+  group     = ibm_is_security_group.fe-ky.id
   direction = "inbound"
-  remote    = proxy-ky.id
+  remote    = ibm_is_security_group.proxy-ky.id
   tcp {
   }
 }
 # Internal. required-connections[2]: (instance fe-ky)->(instance be-ky); allowed-protocols[0]
 resource "ibm_is_security_group_rule" "sgrule-fe-ky-1" {
-  group     = fe-ky.id
+  group     = ibm_is_security_group.fe-ky.id
   direction = "outbound"
-  remote    = be-ky.id
+  remote    = ibm_is_security_group.be-ky.id
   tcp {
   }
 }
 # Internal. required-connections[3]: (instance be-ky)->(instance opa-ky); allowed-protocols[0]
 resource "ibm_is_security_group_rule" "sgrule-opa-ky-0" {
-  group     = opa-ky.id
+  group     = ibm_is_security_group.opa-ky.id
   direction = "inbound"
-  remote    = be-ky.id
+  remote    = ibm_is_security_group.be-ky.id
   tcp {
   }
 }
 # External. required-connections[0]: (external public internet)->(instance proxy-ky); allowed-protocols[0]
 resource "ibm_is_security_group_rule" "sgrule-proxy-ky-0" {
-  group     = proxy-ky.id
+  group     = ibm_is_security_group.proxy-ky.id
   direction = "inbound"
   remote    = "0.0.0.0/0"
 }
 # Internal. required-connections[1]: (instance proxy-ky)->(instance fe-ky); allowed-protocols[0]
 resource "ibm_is_security_group_rule" "sgrule-proxy-ky-1" {
-  group     = proxy-ky.id
+  group     = ibm_is_security_group.proxy-ky.id
   direction = "outbound"
-  remote    = fe-ky.id
+  remote    = ibm_is_security_group.fe-ky.id
   tcp {
   }
 }
